Add swipe-PIP-to-home animation support for ...
non-MATCH_PARENT activity

This CL stops to override the animation bounds to main window frame for
swipe-PIP-to-home animation. It's because the transition moves and
resizes the whole task and shows the launcher. The whole task should
include the partially visible activity behind as well.

This CL also workarounds the animation to restore activity back from
swipe-PIP-to-home. Ideally we should use the whole task leash as
the swipe-to-home animation, but the PIP activity is reparanted
after the animation in PIP1 module. Thus we still override the
animation bounds to main window frame.

We should address the issue in PIP2.

Test: manual - swipe-pip-to-home for match and non-match parent activity
               in portrait and landscape mode
Test: atest PipAnimationControllerTest -j
Bug: 375977163
Bug: 356277166
Flag: com.android.window.flags.better_support_non_match_parent_activity
Change-Id: Ibab51c7606c21b8c199f9c54a81759747a93a5fd