dynamic link to "aconfig_settingslib_flags_java_lib"
"aconfig_settingslib_flags_java_lib" has been added in the framework jar, so the framework should be the source of true of this flag.

statically linked to this flag library will cause different versions of the flag library, and at the runtime, if the wrong version is picked, it will crash the device.

Change-Id: I65788eeaa4056bff7e28d91114eeb6a242d9009e
Test: presubmit
Bug: 378409134
Flag: NONE build file change