[AAPM] Memory Tagging Extension hook
Implement an AdvancedProtectionHook class for turning on Memory Tagging
Extension (as a protection against memory safety issues).

Introduce a hidden API in the DevicePolicyManager to allow controlling
the MTE policy by system components, so that the policy is controlled
only from the DevicePolicyManager (instead of having a circular
dependency between the DPM and AAPM).

Bug: 378931989
Test: atest CtsSecurityTestCases:android.security.cts.advancedprotection.MemoryTaggingExtensionTest
Flag: android.security.aapm_feature_memory_tagging_extension
Change-Id: Ib33baccc20ae36431da9d6944d5a4f0a3233a162