[ADI][UserConfirmation][2/4] Implement new dialogs in Pia to handle incomplete and failed verification
In certain types of failures and the type of installer (priv or
non-priv), the user needs to be shown a dialog to get their response for
continuing or aborting an install. This change implements those dialogs
in Pia.

Bug: 360130528
Test: TBD
Flag: android.content.pm.verification_service
Merged-In: If8b12ee8db04e686a00d8a1389eb3a201db18747
Change-Id: If8b12ee8db04e686a00d8a1389eb3a201db18747