[Notif redesign] Bigger "small" icon in expanded notifs
Updated the remaining expanded templates to use the new header with the
larger icon. We still need to adjust the vertical alignment of the
header though.

There's a KI where the avatars in the old style messaging notification
overlap the app icon, but that will be fixed in a follow-up CL.

Note that:
- notification_2025_template_expanded_big_picture is a fork of
  notification_template_material_big_picture
- notification_2025_template_expanded_big_text is a fork of
  notification_template_material_big_text
- notification_2025_template_expanded_inbox is a fork of
  notification_template_material_inbox
- notification_2025_template_expanded_media is a fork of
  notification_template_material_big_media
- notification_2025_template_expanded_messaging is a fork of
  notification_template_material_big_messaging
- notification_2025_template_expanded_progress is a fork of
  notification_template_material_progress

Bug: 378660052
Test: visual test, screenshot tests to come later
Flag: android.app.notifications_redesign_templates
Change-Id: Iea8327e285d1cfab57fdadf737b544ba433ee726