Add a toolbar for dual shade QS actions
This adds a toolbar for dual shade QS containing:
* Edit mode button
* User switcher button
* Settings button
* Power menu button

It leverages the FooterActionsInteractor as well as the individual view
models that FooterActions uses. This means that it uses the same color
tokens. In the future, we'll need to use material color tokens.

It does not include security "footer", as that will be done later.

Test: manual, test buttons in QS shade
Bug: 381279450
Flag: com.android.systemui.dual_shade
Change-Id: I0ef0ecaa76c1058cff443b5e93f71d9edaf0654a