Override the scrims alpha across the board to match the new UI specs
This is a temporary change that will be cleaned up as part of subsequent CLs.
Without this, surface blur effects are not visible behind the bouncer.
The plan is to move this to ScrimController and rely on scrimAlpha flows provided by different bouncer transitions.

Bug: 381263600
Flag: com.android.systemui.bouncer_ui_revamp
Test: verified manually, the surface behind lockscreen is visible when bouncer is opened.
Change-Id: I27f564268eaf501e3dfcacd7d45e4b1604c2577c